move note components to ui
