+++ modules/periodo-ui/src/index.js
@@ -29,6 +29,7 @@ addUIModules([
   require('./InputBlock'),
   require('./LabeledMap'),
   require('./Links'),
+  require('./Notes'),
   require('./Pager'),
   require('./Patch'),
   require('./Period'),
